backend/requirements.txt
fastapi
uvicorn
pydantic-settings
pydantic
langchain-community
langchain-openaic
langchain-elasticsearch
llama_index
autogen
crewai
psycopg2-binary
python-jose
redis
boto3
elasticsearch
unstructured
sqlalchemy
alembic
python-dotenv
pydantic
asyncpg
pytest
httpx
